put Pool into its own module
not in the cabal file yet (too many other unstaged changes to sort out)
